backend/requirements.txt
fastapi>=0.68.0
sqlalchemy>=1.4.23
pydantic>=1.9.0
uvicorn>=0.15.0
python-dotenv>=0.19.1
aiohttp>=3.8.1
pytest>=7.0.0
httpx>=0.28.1 
redis>=5.0.8
